requirements.txt
kafka-python==2.0.2
pyspark==3.5.1
findspark==2.0.1
python-dotenv==1.0.1
requests==2.32.0
